+++ esp32/snapclient/include/snapclient_mode.h
@@ -9,7 +9,9 @@
 
 #include "audio_probe_stream.h"
 #include "audio_output_controller.h"
+#include "project_snap_output.h"
 #include "snapcast_pcm_decoder.h"
+#include "snapclient_time_sync.h"
 #include "runtime_mode.h"
 
 class ProjectSnapProcessorRTOS;
@@ -25,13 +27,19 @@ class SnapclientMode : public RuntimeMode {
 
  private:
   bool connectWifiWithTimeout();
+  static void snapClientTaskEntry(void *context);
+  void snapClientTaskLoop();
 
   WiFiClient wifiClient_;
   AudioOutputController audioOutput_;
   AudioProbeStream pcmProbe_;
   SnapcastPcmDecoder codec_;
+  ProjectSnapOutput snapOutput_;
   std::unique_ptr<ProjectSnapProcessorRTOS> snapProcessor_;
   snap_arduino::SnapClient snapClient_;
+  SnapTimeSyncClampedDynamicSinceStart dynamicTimeSync_;
+  TaskHandle_t snapTaskHandle_ = nullptr;
+  volatile bool snapTaskRunning_ = false;
   uint32_t lastWifiCheckMs_ = 0;
   bool playbackIdleLogged_ = false;
 };